Import 2.3.20pre1

+++ arch/ppc/kernel/indirect_pci.c
@@ -19,7 +19,7 @@ unsigned char * pci_config_data;
 int indirect_pcibios_read_config_byte(unsigned char bus, unsigned char dev_fn,
 			     unsigned char offset, unsigned char *val)
 {
-	unsigned flags;
+	unsigned long flags;
 
 	save_flags(flags); cli();
 	
@@ -35,7 +35,7 @@ int indirect_pcibios_read_config_byte(unsigned char bus, unsigned char dev_fn,
 int indirect_pcibios_read_config_word(unsigned char bus, unsigned char dev_fn,
 			     unsigned char offset, unsigned short *val)
 {
-	unsigned flags;
+	unsigned long flags;
 	
 	if (offset&1) return PCIBIOS_BAD_REGISTER_NUMBER;
 
@@ -53,7 +53,7 @@ int indirect_pcibios_read_config_word(unsigned char bus, unsigned char dev_fn,
 int indirect_pcibios_read_config_dword(unsigned char bus, unsigned char dev_fn,
 			     unsigned char offset, unsigned int *val)
 {
-	unsigned flags;
+	unsigned long flags;
 	
 	if (offset&3) return PCIBIOS_BAD_REGISTER_NUMBER;
 
@@ -71,7 +71,7 @@ int indirect_pcibios_read_config_dword(unsigned char bus, unsigned char dev_fn,
 int indirect_pcibios_write_config_byte(unsigned char bus, unsigned char dev_fn,
 			     unsigned char offset, unsigned char val)
 {
-	unsigned flags;
+	unsigned long flags;
 
 	save_flags(flags); cli();
 	
@@ -87,7 +87,7 @@ int indirect_pcibios_write_config_byte(unsigned char bus, unsigned char dev_fn,
 int indirect_pcibios_write_config_word(unsigned char bus, unsigned char dev_fn,
 			     unsigned char offset, unsigned short val)
 {
-	unsigned flags;
+	unsigned long flags;
 
 	if (offset&1) return PCIBIOS_BAD_REGISTER_NUMBER;
 
@@ -105,7 +105,7 @@ int indirect_pcibios_write_config_word(unsigned char bus, unsigned char dev_fn,
 int indirect_pcibios_write_config_dword(unsigned char bus, unsigned char dev_fn,
 			     unsigned char offset, unsigned int val)
 {
-	unsigned flags;
+	unsigned long flags;
 
 	if (offset&3) return PCIBIOS_BAD_REGISTER_NUMBER;
